refactor for clean ui
